+++ packages/cdk/lib/types.ts
@@ -28,6 +28,7 @@ export interface CubeCobraStackParams {
   draftmancerApiKey: string;
   stripeSecretKey: string;
   stripePublicKey: string;
+  enableBotSecurity: boolean;
 }
 
 export type Environment = 'production' | 'development' | 'local';