+++ CHANGES.md
@@ -8,5 +8,10 @@ As of 1.4.57
 
 - Fixed an issue where paginating cube search results would sometimes not go to the next page
 - Fixed a bug where updating user fields would delete password
+- Fixed a bug where seat owner names were not showing up
+- Fixed a bug with RSS feeds erroring
 
 # Technical Changes
+
+- Added codepipeline with integration tests
+- Fixed docker development setup